obtener alumnos 2/requirements.txt
functions-framework==3.*
google-cloud-bigquery
requests
datetime
pandas
pyarrow
google-cloud-pubsub>=2.6.1